More Refactoring
Really aggressively moving things around, hopefully it should result in
an easier to maintain project with greater scope for extensibility and
other hardware vendors.
