+++ KSP-LogiRGB/KSP-LogiRGB/KSP-LogiRGB.csproj
@@ -61,10 +61,10 @@
     <Compile Include="ColorSchemes\LogoScheme.cs" />
     <Compile Include="ColorSchemes\VabScheme.cs" />
     <Compile Include="Config.cs" />
-    <Compile Include="IDataDrain.cs" />
     <Compile Include="KSPLogiRGBPlugin.cs" />
-    <Compile Include="Logitech\LEDControl.cs" />
-    <Compile Include="Logitech\LogitechDrain.cs" />
+    <Compile Include="LEDControllers\ILEDController.cs" />
+    <Compile Include="LEDControllers\Logitech\LEDControl.cs" />
+    <Compile Include="LEDControllers\Logitech\LogitechLEDController.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SceneManagers\FlightSceneManager.cs" />
     <Compile Include="SceneManagers\SceneManager.cs" />
